+++ app/webroot/src/sass/generate_sprites.scss
