+++ app/webroot/src/sass/main.scss
@@ -1,5 +1,5 @@
 @import "compass";
 
-@import "sprites";
+//@import "sprites";
 
 @import "styles";
